[Implementation] Implemented Character
Created the Character class, which is a RectangleShape that moves with wasd.
Currently Character extends Platform, this is weird, I dont plan on keeping it this way
I may end up making an Object interface for things with an update() method,
or rename Platform to something more generic.

For now, this works, but it will be changed once I'm actually trying to make
a more versitile object management system.