[Implementation] Implemented some simple collision detection
The main loop now checks for collision after updated all the objects, then
moves the character upward if it is colliding with any of the platforms.

Character was also modified so it always moves down if you aren't moving up,
creating a kind of gravity that allows you to land and walk on the platforms.

Collision with the sides of the platforms is a little wacky.